Update BedSearchRepository to account for the new overlap member.
Associated tests for these have also been updated. During testing a small mistake was found in the instantiation of various bookings. Included in this commit fixes this mistake, namely changing the arrivalDate and departureDate of bookings with specific temporal descriptors on associated beds to their correct values. For example, the instantiated bed  bedThreeInRoomOneInPremisesOneMatchingEverything was given the name "Matching bed with overlapping booking in middle", but its associated booking contained arrivalDate and departureDate values which indicated an overlapping booking spanning the end of the date range provided in BedSearchRepository.findTemporaryAccommodationBeds, not the middle.