Increase memory size for Community API responses
We are getting some absolutely huge responses (400k+) from Community
API for some CRNs which is blowing out the default in-memory size that
Spring expects (262144 bytes). This adds a config value to increase this
limit to try and combat this. Going forward, we’ll want to optimise the
response and there is an ongoing discussion with the integrations team
about this, but upping the memory limit should hopefully fix the
immediate problem.