Attempt to fix offender name update task
This seemed to work in every environment except from prod. My assumption
is that it’s because there’s more data in prod, and some applications
have already been updated with names, so something’s getting out of
whack. With this in mind I’ve updated the task to only fetch
applications with Null names. While testing, this seemed to effect
pagination too, so I’ve updated the pagination query to return a `Slice`
(https://docs.spring.io/spring-data/commons/docs/current/api/org/springframework/data/domain/Slice.html)
which only knows if there is a nextPage. After much wailing and gnashing
of teeth, we realised that we only need to get the first page of results
each time, as when we update the results, this does not update the
counter, leading every other page to get skipped.