Updated OpenApi specification to relegate `BookingSearchResultPersonSummary.name` to be an optional field.
This is required for allowing a person's name to be `null` - previously the field was required and therefore unnamed individuals were set to an empty string value. We require a person's name to be null in LAO cases where we should not reveal their name to the current logged-in user.

Associated transformer function and test involving `BookingSearchResultPersonSummary.name` has been changed to reflect this change.