create an `oasys/risk-to-self` endpoint
CAS2 would like to request only the Risk to Self section of an OASys.
(ministryofjustice/hmpps-community-accommodation-tier-2-ui#176 )

Currently there is the `oasys/sections` endpoint which enforces returning
several sections (e.g. ROSH, Risk to Self, Risk Management Plan) and
optionally returns some other sections. It also does not return one of
the questions we need in Risk to Self ('historical risk of self-harm')

This PR suggests an `oasys/risk-to-self` endpoint which only returns the
Risk to Self section.