Assign default role of `CAS3_ASSESSOR` for users in Temporary Accommodation
There are over 13,000 CPPs as potential users of the Temporary Accommodation
service, and collecting the information needed to seed everyone into the
database is a prohibitively difficult task.
It's therefore easier to seed all of the users for whom some other role is
appropriate and allow the Temporary Accommodation service to assign the
`CAS3_REFERRER` role to anyone who doesn't already have a CAS3 role assigned
to them when they sign in for the first time.