Accept summary data as a parameter to `AssessmentService.createTemporaryAccommodationAssessment`
As Temporary Accommodation assessments now require some summary data
to be provided, the general `createAssessment` function will no longer
be useful as Approved Premises and Temporary Accommodation will need
different sets of parameters to construct an assessment.

This commit makes the two specific functions public and updates
`createTemporaryAccommodationAssessment` to serialize the summary data
as JSON and store it on the assessment.