Add an ApplicationReportGenerator
This adds a new generator for Application reports, as well
as the associated Row and Property classes. We override the
`filter` function to just return true, as we’ll be doing the
filtering in the SQL side. Some of these fields return null at the
moment, as we don’t know how to fetch the data (yet), but can iterate
as we go.