Ensure name search works correctly
Before the name search wasn’t working as expected, as if `crnOrName`
was provided, it was passing the that variable as the CRN too, meaning
the CRN query trumped the name query, so the name query never worked.
This moves the boolean logic to the query itself, so we can handle names
and CRNs correctly. Going forward, we’ll remove the CRN query, but we’ll
keep it around until the frontend is updated.