Increase Case Notes timeout to 30 seconds
We’re having some issues with certain case note responses timing out.
I’ve added a special timeout value for case notes so we can hopefully
fix this. Long term fix is to get a custom API endpoint for this kind
of info, but this should fix things in the short-term.