Fix assessment state transition from 'ready to place' back to 'in review'
When trying to mark an assessment that is 'ready to place' as being
'in review' again, the API would not correctly update the state, causing
the assessment to still be shown as 'in review'.

This was caused by the `AssessmentService.reallocateTemporaryAccommodationAssessment`
method not resetting the decision on the assessment to `null`.
Making this change results in the expected behaviour.

The `AssessmentStateTest` class contains a test method that verifies
all of the expected state transitions work as intended.