Ensure seed CSVs are run in order
This appends numbers in front of each CSV to ensure they
run in the correct order. We then remove the number to ensure the
correct job type is run. I’ve also updated the tests to verify that
the jobs are run in the correct order. I decided to use a spy to
make sure that the `seedData` method is called. This is tested in
the integration tests anyway, so we don’t lose any coverage.