Add a `notApplicable` releaseType option
If an application has a non statutory sentence type, then currently, we
cannot submit an application, as there is not necessarity a relevant
release type for this type of sentence. This is a very unusual
circumstance, as non-statutory sentence types would not ordinarily be
eligible for AP, but this happens often enough for us to have to handle
it.

After some discussion, we decided to add a `notApplicable` option,
rather than making it non-nullable, which would have ramifications
elsewhere.