Return a 404 when a user cannot be found
As we’re reusing `getUserForUsername`, this adds a `throwProblemOn404`
option, which defaults to `false`. If it’s set to `true` and the
response from Delius is a 404, we throw a `NotFoundProblem`, which
bubbles up to the controller as a 404. I’ve also tweaked the
`NotFoundProblem` to accept an optional `identifier`, which allows us to
customise the JSON body of the 404 response if we need to.