Add pagination to user list
This adds pagination to the users endpoint, which allows us to paginate
and sort the response, if the `page` argument is added. I’ve also added
some util functions that should make dealing with pagination a bit
easier / more DRY in future. These functions are now being used in the
`getAllActive` function in the PlacementRequestService too.