Correct regional mapping code for Greater Manchester
There was work 3 months ago to introduce a mapping layer ontop of probation regions called 'probation_area_probation_region_mappings’.[1]

Our latest attempt to seed an HPT user failed with this error related to incorrect probation region for N50[2]. This is odd because we’ve seeded users with N50 before and they’re using the service.

This appears to have copied over the wrong Delius region code for Greater Manchester[3]. Instead of MCG it should be N50. I suspect the latest work missed the migration we made to correct this code[4] and took the values from a seed file rather than from the real values.

[1] https://github.com/ministryofjustice/hmpps-approved-premises-api/commit/a1551b3d4daa9f749cfbbcd3fdca972f72387762
[2] https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/main/src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/service/UserService.kt#L262
[3] https://github.com/ministryofjustice/hmpps-approved-premises-api/pull/758/commits/e01822dde4c60fa2ab952cad78ff8fc03991e8b5#diff-0bd31fc8ee0102525db3c54da3ecb6e42826cf10621cbf5f36ec5acfe7fbafdaR14
[4] https://github.com/ministryofjustice/hmpps-approved-premises-api/blob/e6fdda8201262c776a7c9b05ae6c7b9bf0914e57/src/main/resources/db/migration/all/20230214123554__correct_delius_code_for_greater_manchester_region.sql#L2