Get premises type correctly in Application reports
At the moment, the premises type is returning an integer, which is what
gets stored in the databse, for example, if the premises type is
`normal`, we get `0`, premises type is `pipe` we get `1` and so on. This
updates the query to return the `ap_type` as `premisesTypeIndex`. We
can then change the `getPremisesType` interface method to get the index
(if it exists) and use `ApType.entries` to get the enum value by the
index. I suspect this is how it’s done with normal JPA classes, but with
query interfaces, it seems we need to jump through this additional hoop!