Add offence ID to Temporary Accommodation applications
Currently there is no way to retrieve the index offence for a referral
in the Temporary Accommodation service.
When creating an application, the offence ID is required and stored in
the database already.
By including the offence ID in the `TemporaryAccommodationApplication`
schema, it becomes accessible when users of the API send a request
for an application or for an assessment (which include its application),
allowing the presentation of the index offence.