Add integration test for booking with missing inmate details
This integration test verifies that the `GET /premises/{premisesId}/bookings`
endpoint will return successfully when the Prison API cache contains an
unsuccessful response (such as a 500 Internal Server Error). To enable this
test, the `Given an Offender` utility method has been modified so that the
Wiremock calls to the Community API or the Prison API can be configured to
return a 500 response using the
`mockServerErrorForCommunityApi`/`mockServerErrorForPrisonApi` parameters.