Update `OffenderService.getInmateDetailByNomsNumber` to not throw an exception when the cache fails
To prevent an unsuccessful response from the Prisons API (from the cache or
otherwise) causing a 500 response to be returned from the API, this method has
been updated not to explicitly throw an exception. In situations where an
exception would have been thrown, the exception is still generated, but logged
instead at the `WARN` level and `null` is returned instead. Additionally, the
preamble to the warning log message specifies whether the error happened after a
cache timeout or if the cache itself contains the error response.