Hide reallocated placmement request in task list
I was wracking my brains about why there were so many placement requests
that appeared to be out of date in the task list. While all the other
task queries are JPA derived queries (i.e `findBy..`), the Placement
Request query is a custom one, and we left out the
`reallocated_at IS NULL` conditional. This should fix it and mean there
are much fewer tasks in the list now.