Don’t fail when unable to allocate a user
We’re getting an unusual number of allocations failing when submitting
applications etc. While we attempt to work out what’s going on, we’ll
log errors and make the allocated to user null. We’ll also log the
failure to allocate together with the number of attempts, which will
give us more of an idea why these errors are happening.

As part of this we’ve had to make the allocated to user nullable in a
few places.