Ensure qualified users don’t get asssigned
In a previous commit, we removed the check for if a potential users
has a qualification and the application requires no qualification. This
is needed to ensure that (in particular) Pipe and Esap assessors don’t
get assigned standard applications (something they’re not qualified to
do). Rather than doing this check in the code, we update the query to
add a `qualifiedUserRequired` switch. If this switch is set to zero,
then we add an additional query to omit any users with any Non-LAO
qualifications.