Add an AllocationsEngine class
The Allocations engine has been giving us a lot of grief recently. I
attempted a fix in a previous PR, but I’ve since discovered that it
didn’t _quite_ match up with what was expected. To try and make this
a bit more configurable, and to remove some of the complex logic, I’ve
added a `UserAllocationsEngine` class, which has a `getUserPool`
function that uses the `CriteriaBuilder` to get a list of qualified
users IDs, we then pass those IDs to get the user with the smallest
workload, depending on the type of task.

It’s still a little cumbersome, but I’ve attempted to make it a little
easier to grok with comments etc.