Persist targetLocation in applications
This field is already passed when submitting an application, but only
gets used in domain events. This adds a new targetLocation field to the
application location, so we can use it in reporting. We will need to
look at backfilling the applications that do not have this data,
possibly using the domain events.