updateUserFromCommunityApiById will now also update the user’s region
CAS3 users are instructed to update their region if they are unable to use the service. When they do so, our users table and the probation_region_id are not updated. In order to automate this we hook into this existing class which is invoked by a job to update the region if a new region mapping can be found.