Add built specs to repo to allow TS type generation
We need these build artefacts in version control in order to
generate Typescript types in the UI projects.

See the [`generate_types` script] in the CAS2 UI project for
example. This will be tweaked to use the generated `built-cas2-api-spec.yml`
file as this process is also unable to follow `$ref` links.

REFERENCES

- [`generate_types` script]: https://github.com/ministryofjustice/hmpps-community-accommodation-tier-2-ui/blob/main/script/generate-types