Add a generator for daily metrics
This is a little bit more fiddly than other metrics generators, as we
need a row for each date. To get around this, we pass a list of domain
events and applications to the class construtor and then `createReport`
function accepts an array of dates, we then loop through each date, and
filter the applications and domain events accordingly.

To get around having to use magic numbers I’ve set the tests up to
build a Map of Keys, which are mapped to dates and unique users. That
way we can count all the mapped values for a particular date to get the
total number of events, and then a count of the keys for a particular
date gives us the total number of unique users.