Set OAuth2 client registration correctly
The OAuth client used to authenticate the system-to-system connection
between the CAS and the new Nomis-User-Roles client must be configured
with a key in `security.oauth2.client.registration` in `application.yml`
which matches the environment variable defined in `helm_deploy/values.yaml`
in the form `SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_*`, in this case:

  - `nomis-user-roles-api` in `application.yml` and
  - `SPRING_SECURITY_OAUTH2_CLIENT_REGISTRATION_NOMIS-USER-ROLES-API_CLIENT-{ID,SECRET}`
     in `values.yaml`

Additionally, we need to tell our API client which OAuth client
to use. This is done in `WebClientConfiguration.kt` using the
value set in `security.oauth2.client.registration` in `application.yml`,
in this case `"nomis-user-roles-api"`.