Allow metrics to be grouped by AP Type
This adds a generic and some additional logic to allow us to group our
rows by AP Type as well as Tier. I’ve also moved the `Tier` class to the
models folder and given it a more specific name, alongside a new
`ApTypeCategory` enum. I’ve also updated the `tier` variable in the
Report row to be more generic.