Allow CAS2_ASSESSOR role to GET Cas2Applications
`GET` requests for either a list of applications or an
individual application are now permitted to the
`CAS2_ASSESSOR` role in addition to the `PRISON` role.

i.e. our external NACRO users, who have the CAS2_ASSESSOR
role are granted access to list and view CAS2 applications.
Internal users with the PRISON role continue to have that
access.