add environment variables for CAS2 domain events
We would like to enable a basic CAS2 domain events in the dev environment, so that we can start testing them and understanding what types of data and events we need to emit.

This commit adds the values needed for the url template, and a feature flag to enable them in dev and disable emitting the events in test, preprod and prod.