add CAS2 domain events Open API
We would like to enable a basic CAS2 domain events in the dev environment, so that we can start testing them and understanding what types of data and events we need to emit.

This commit adds endpoints needed to get a basic 'application submitted' event.

It includes a new type of Staff Member, because of the way Nomis Users are structured.
It also includes a slightly different Person Reference, because CAS2 is prison-based only the noms number is required (afaik).