Add a generator for Placement Metrics
This takes a list of `ApplicationTimelinessEntity`s checks how many
applications resulted in a successful booking, as well as the average
time in days it took for bookings to be made, both from the point of
application and the point of assessment. We also use the
`workingDayCountService` to count the timeliness in working days, and
break down the number of applications by time window.

To make testing easier in the unit tests, we mock the
workingDayCountService to just count the difference in calendar days.