CAS2 Domain Event Service
We would like to enable a basic CAS2 domain events in the dev environment, so that we can start testing them and understanding what types of data and events we need to emit.

This creates a basic service, modelled on the CAS3 one, to get a CAS2 domain event, and save and emit an ApplicationSubmitted event.