Save and emit CAS2 event on Application submitted
We would like to enable a basic CAS2 domain events in the dev environment, so that we can start testing them and understanding what types of data and events we need to emit.

This calls the previously created CAS2 Domain Event Service to save and emit an event when a CAS2 application is submitted.