Return all only submitted applications to assessor
GET /cas2/applications

A note on ApplicationStatus versus 'has been submitted'
---

Note that for "submitted" we test for the presence of a
value in the `Cas2Application.submittedAt` timestamp. An
application can be submitted exactly once and can never
become "unsubmitted".

i.e. "submitted" is an event rather than a status which
changes from time to time. This is contrary to what is
currently implied by ApplicationStatus:

```
ApplicationStatus:
  type: string
  enum:
    - inProgress
    - submitted
    - requestedFurtherInformation
    - pending
    - rejected
    - awaitingPlacement
    - placed
    - inapplicable
    - withdrawn
```

In CAS2 we will be requiring external users from NACRO
(logging in with the CAS2_ASSESSOR role) to provide updates on
their progress in taking submitted applications through the
assessment and placement process which they manage in their
own NACRO / OpenHousing systems.

So it will be clearer for CAS2 to model something
like a `CAS2ExternalWorkflowStatus` which can go though
states such as:

- `pendingAssessment`
- `furtherInfoRequested`
- `applicationWithdrawn`
- `accepted / rejected`
- `placed`
- `bookingCancelled`

We probably need to implement a lightweight state-machine to
ensure that only valid transitions are made. i.e. that an
application isn't moved from `pendingAssessment` to
`bookingCanclled` without having been first `accepted`.

As these workflow updates will be made by NACRO outside of
their actual workflow perhaps it will be necessart for us to
ask for the date on which the status change occurred. e.g.
"when was this application accepted?" as this is not
necessarily when the event is going to be reported on our
service...