Add /cas2/submissons endpoints for GETting submitted application(s)
When the UI asks that API for a list of applications, what is the appropriate
range of filtering which should be implicitly applied by the API?

`GET /applications`
  - APPLICANT: all applications which I've created (filter on associated user)
  - ASSESSOR: all application which have been submitted (filter on submittedAt
    presence)
  - ADMIN: all applications (no filter)

`GET /application/{applicationId}`
  - APPLICANT: 403 if I didn't create the application
  - ASSESSOR: 404 if the application hasn't yet been submitted

I think it's reasonable to implicitly filter a list of resources according to
some access rules, which could become slightly complex, such as:

- APPLICANT can see applications they've created
- APPLICANT can see applications made to their region
- APPLICANT can see applications where an assessor has named them in a request
  for further information

It becomes more opaque if we combine this with rules which are based on the
**state of the application** as well as on the **role of the user**, e.g.

- ASSESSOR can see only applications which have been submitted

The submission of an application is an important milestone (consider the
`application.submitted` domain event). The mutable form `data` is stored in a
permanent read-only `document` as a full and reliable record of the risks and
needs at the time of application.

The lifecycle of the accommodation workflow moves to a new stage on submission
of the application. It's not a  stage out which the application can transition,
it's an event which cannot be undone.

If we agree that the API should not handle filtering submitted applications
implicitly, how should the UI request them explicitly? Two options:

1. url params, e.g. `GET /cas2/applications?filters=[submitted]`. In this
   approach the complexity at the API is retained: the controller actions still
   needs to handle the logic of whether to return all applications or just those
   which have been submitted.
2. dedicated url e.g. `GET /cas2/submissions`. In this approach a new endpoint
   and controller (and API representation) can be used.

I believe that by giving this different entity its own name e.g. **submission**
we can make our code simpler.

The new `cas2/submissions` endpoint can return a new representation of an
application which has been through the submission process. This
SubmittedApplication or SubmittedApplicationSummary can do a much better (more
specialised) job of meeting our needs as we leave the "apply" journey behind.
e.g.

- `data` is no longer relevant or helpful. Only exposing `document` is clearer
  and less error prone in displaying a submitted application.
- we can add in a list of `StatusUpdates` to our new `SubmittedApplication`
  representations