Create domain event for manual booking
This updates the code to create a domain event regardless of if the
associated application is an offline application or not.

The only circumstance where we DON’T want to create a domain event
is when we’re seeding bookings. Rather than just checking for the
presence of a user before creating a domain event, we add a variable
called `isCalledFromSeeder`, which gives more idea of intent. Also,
when seeding, we don’t want to check for the presence of a eventNumber

I’ve also tidied up the tests for `createApprovedPremisesAdHocBooking`
to DRY up the code and make the tests tidier and more readable.