Skip OpenAPI Generate task correctly
We recently added some config to skip the OpenAPI generation task when
running tests in IntelliJ, to try and speed up testing in development.
However, before, we were skipping the OpenAPI generation task in the
withType block directly.

Weirdly, because of the Grade build lifecycle, doing it this way meant
that we skipped the tasks for every task that was run, not just tests.
This meant that if we tried to run the openApiGenerate task within
Intellij, the task got skipped and just didn’t run.

This adds an `onlyIf` conditional to skip all the OpenAPI generation
tasks if the dependent task is a test type task and the task is being
run from inside IntelliJ