Install Detekt for static code analysis
We asked hmpps_dev if there was a standard choice given deteck and sonarqube are referenced across MoJ Slack. The loose suggestion[1] was to use detekt.

We have to do a bit of configuration magic to silence a spring compatibility error. We copy this other example of a project installing this library[2].

[1] https://mojdt.slack.com/archives/C69NWE339/p1698668913072199
[2] https://github.com/ministryofjustice/hmpps-allocations/pull/8/commits/aaa202d5bdf70e93324c99e6bc6282df20b96eae