CI checks static code analyser for new pull requests
This behaves like unit_tests and integration_tests in that it’s not required for the deployment. Presumably because these tasks are always run before a change can be merged we can trust that they pass.

I thought to include `with_detekt` in the name so it makes the tool we use more discoverable.