Create performance test for generating a bookings report
This performance test introduces around 1200 bookings with arrival dates
randomly distributed around a 120-day period centred on the current date,
and then generates a sample of around 100 reports for a random month within
that window of time. This results in, on average, around 300 bookings
per report.

Given the current approach of hitting the offender details cache for
every unique CRN in the list of bookings, the report takes over 60
seconds (and the connection times out) around 40% of the time. A further
30% of the time the API returns a 500 response as the database
connection pool is drained, but this is an artifact of the number of
concurrent users generating reports to complete the test in a timely
manner, as it is unlikely that 100 users would all request booking
reports in a less than 10 minute timespan.