Use bulk personal data endpoint to construct personal data columns
The original approach has unacceptably poor performance for realistic
numbers of rows on the bookings report. Using the bulk personal data
endpoint improves this significantly at the cost of not being able to
include all the desired information (notably, the PNC number is not yet
provided by this endpoint). The missing information is retained on the
spreadsheet, but always empty.