Abandon premptive caching for records that we fail 25 times on
There are a handful of cases where NOMS numbers cannot be found upstream. eg: a6833DY and A0943AE[2].

We confirmed that these NOMS numbers have been ‘merged’[3] which means they will never return data again. Since we have referrals or bookings that reference these noms numbers, the service and it’s premptive caching logic will attempt to cache offender data for these people.

As a way to handle this situation we introduce an upper limit for failure. Currently set to 25 which includes back offs. If the cache warmer cannot find information in 25 requests we give up. We mark the metadata as abandoned and skip any futher requests to fetch this information.

Meta data is stored in redis and could theoretically be unset. If that happened new attempts would attempt to recache 25 times before getting back into the same abandoned state.

[1] https://ministryofjustice.sentry.io/issues/4538149828/?project=4503931792392192&query=is%3Aunresolved&referrer=issue-stream&statsPeriod=1h&stream_index=0
[2] https://ministryofjustice.sentry.io/issues/4521955843/?project=4503931792392192&project=4504129156218880&query=is%3Aunresolved&referrer=issue-stream&statsPeriod=1h&stream_index=1
[3] https://mojdt.slack.com/archives/CR5ESQ8T1/p1694695349149839