Implement auth controls and a 'noop' happy path
Only authenticated users with the `CAS2_ASSESSOR` role are
permitted to create status updates.

Our controller function does nothing for now, just returns a
201 Created response code.

Due to the limitations of the Kotlin Spring OpenAPI code
generation, the controller function to handle this request
needs to live in the `SubmissionsController` (the first
portion of the url scheme, in this case /submissions/ is
mapped to the controller name). It would
probably be clearer and more maintainable to use a separate
controller (e.g. `StatusUpdatesController`), but at present
this won't add much additional length or complexity to the
SubmissionsController:

- one new function to handle the POST
- the inclusion of a new service to create and retrieve
  status updates (coming soon)