Increase maximum database connections to 20 per replica
There are many examples of other teams increasing this pool size:

- hmpps tier to 25 https://github.com/ministryofjustice/hmpps-tier/blob/8faa4bfd31580767bed8bba453cdddd70098f3ad/src/main/resources/application.yml#L17
- csr-api to 35 https://github.com/ministryofjustice/csr-api/blob/536fc26119c38d0bc4c83f2ad52482bb39414e00/src/main/resources/application.yml#L13
- community-api to 30 https://github.com/ministryofjustice/community-api/blob/54fdf94c9ce2e6e534ac18c11c82c73ff1f2995d/src/main/resources/application.yml#L9

Tier in particular increases to 25 whilst they have a db.t4g.small instance in prod[3]. They do only have 2 replicas in prod[4] whilst we have 4, any value we set will result in double the connections to our database.

Our RDS instance in prod is a db.t3.small[2]

When we query on preprod we see the number of connections for this instance is 192:

```
show max_connections;
-- success --
max_connections
---------------
192
```

If we set this to 20 and * 4 replicas our new maximum should be 80, up from 40. Well within the database’s max of 192.

[2] https://github.com/ministryofjustice/cloud-platform-environments/blob/main/namespaces/live.cloud-platform.service.justice.gov.uk/hmpps-community-accommodation-prod/resources/rds.tf#L13C35-L13C46
[3] https://github.com/ministryofjustice/cloud-platform-environments/blob/main/namespaces/live.cloud-platform.service.justice.gov.uk/hmpps-tier-prod/resources/rds.tf#L16
[4] https://github.com/ministryofjustice/hmpps-tier/blob/8faa4bfd31580767bed8bba453cdddd70098f3ad/helm_deploy/hmpps-tier/values.yaml#L4C3-L4C15