Create Cas2StatusUpdate when POSTing valid status
In this commit we:

Define a new `Cas2StatusUpdateEntity`
---

These entities will record the individual status updates for a
particular application. Each status update include:

- association with an application
- association with an assessor
- association with a status option

we also record:

- the label of the chosen status option, and
- the description of the chosen status option

and we believe that these are liable to change over time. The status itself will
not be allowed to change. For example:

- if the label of `awaitingDecision` changes from `Awaiting decision` to `Pending decision`
  we will record these different labels in records with the same `statusId`. However,

- if a replacement status called `pendingDecision` is required, we will:
  - create a new status with that name
  - mark the `awaitingDecision` as no longer active

Establish some new JPA relationships
---

- a `Cas2Application` now "has many" `Cas2StatusUpdate`s
- a `Cas2StatusUpdate` "belongs to"

  - an "assessor" (External User)
  - an "application" (Cas2Application)

Manually handle the relationship between status updates and the status options
---

The status options available are modelled in `Cas2Application` and are defined in
`Cas2ApplicationStatusSeeding`. As they aren't yet stored in the database  we set
`Cas2StatusUpdateEntity.status_id` manually rather than using the JPA `@ManyToOne`
annotation.